requirements.txt
numpy==1.19.5
tensorflow-gpu==2.4.0
tqdm==4.62.3
streamlit==1.8.1
opencv-python==4.5.3.56
Pillow==8.3.1
sklearn==0.0
scikit-learn==0.20.4